fix buttons in filmModal
