+++ src/library.html
@@ -13,7 +13,7 @@
     </div>
     <include src="./partials/library_bg.html"></include>
     <include src="./partials/galery.html"></include>
-    <!-- <include src="./partials/pagination.html"></include> -->
+    <include src="./partials/pagination.html"></include>
     <include src="./partials/footer.html"></include>
     <include src="./partials/team-modal.html"></include>
     <include src="./partials/global-loader.html"></include>
