+++ src/partials/footer.html
@@ -9,7 +9,7 @@
         <use href="./images/icons.svg#icon-heart"></use>
       </svg>
       by
-      <a href="#" class="footer-link" data-open-modal-team>GoIT Students</a>
+      <button class="footer-link" data-open-modal-team>GoIT Students</button>
     </p>
   </div>
 </footer>
