+++ src/sass/components/footer.scss
@@ -63,6 +63,7 @@
 }
 
 .footer-link {
+  border: none;
   display: inline-block;
   font-family: inherit;
   font-weight: 400;
